Move instantiate_rigids to types crate
