+++ compiler/late_solve/src/lib.rs
@@ -12,10 +12,9 @@ use roc_types::subs::Content;
 use roc_types::subs::{ExposedTypesStorageSubs, Subs, Variable};
 use roc_unify::unify::{unify as unify_unify, Mode, Unified};
 
-pub use roc_solve::solve::instantiate_rigids;
-
 pub use roc_solve::ability::resolve_ability_specialization;
 pub use roc_solve::ability::Resolved;
+pub use roc_types::subs::instantiate_rigids;
 
 #[derive(Debug)]
 pub struct UnificationFailed;
